+++ client/resteasy-client/src/main/java/com/cerner/beadledom/client/resteasy/ResteasyCorrelationIdContext.java
@@ -1,7 +1,6 @@
 package com.cerner.beadledom.client.resteasy;
 
 import com.cerner.beadledom.client.CorrelationIdContext;
-import com.cerner.beadledom.correlation.CorrelationContext;
 import javax.annotation.Nullable;
 import javax.ws.rs.core.HttpHeaders;
 import org.jboss.resteasy.spi.ResteasyProviderFactory;
@@ -20,19 +19,16 @@ import org.slf4j.MDC;
 class ResteasyCorrelationIdContext implements CorrelationIdContext {
   private final String headerName;
   private final String mdcName;
-  private final CorrelationContext proxyCorrelationContext;
 
-  ResteasyCorrelationIdContext(@Nullable String headerName, @Nullable String mdcName,
-      CorrelationContext correlationContext) {
+  ResteasyCorrelationIdContext(@Nullable String headerName, @Nullable String mdcName) {
     this.headerName = headerName != null ? headerName : CorrelationIdContext.DEFAULT_HEADER_NAME;
     this.mdcName = mdcName != null ? mdcName : CorrelationIdContext.DEFAULT_MDC_NAME;
-    this.proxyCorrelationContext = correlationContext;
   }
 
   @Override
   public String getCorrelationId() {
 
-    String correlationId = proxyCorrelationContext.getCorrelationId();
+    String correlationId = com.cerner.beadledom.correlation.CorrelationIdContext.get();
     if (correlationId != null) {
       return correlationId;
     }