+++ jaxrs/src/main/java/com/cerner/beadledom/jaxrs/CorrelationIdFilterProvider.java
@@ -1,6 +1,5 @@
 package com.cerner.beadledom.jaxrs;
 
-import com.cerner.beadledom.correlation.CorrelationContext;
 import com.cerner.beadledom.jaxrs.provider.CorrelationIdFilter;
 import com.cerner.beadledom.jaxrs.provider.CorrelationIdHeader;
 import com.cerner.beadledom.jaxrs.provider.CorrelationIdMdc;
@@ -24,12 +23,8 @@ class CorrelationIdFilterProvider implements Provider<CorrelationIdFilter> {
   @Inject(optional = true)
   String correlationIdMdc;
 
-  @Nullable
-  @Inject(optional = true)
-  CorrelationContext correlationContext;
-
   @Override
   public CorrelationIdFilter get() {
-    return new CorrelationIdFilter(correlationIdHeader, correlationIdMdc, correlationContext);
+    return new CorrelationIdFilter(correlationIdHeader, correlationIdMdc);
   }
 }