+++ jaxrs/src/test/scala/com/cerner/beadledom/jaxrs/provider/CorrelationIdFilterSpec.scala
@@ -1,6 +1,6 @@
 package com.cerner.beadledom.jaxrs.provider
 
-import com.cerner.beadledom.correlation.ThreadLocalCorrelationContext
+import com.cerner.beadledom.correlation.CorrelationIdContext
 import org.scalatest._
 import org.scalatest.mockito.MockitoSugar
 import org.slf4j.MDC
@@ -9,21 +9,20 @@ class CorrelationIdFilterSpec extends FunSpec with BeforeAndAfter with Matchers
 with MockitoSugar with CorrelationIdFilterBehaviors {
   val defaultHeaderName = "Correlation-Id"
   val defaultMdcName = "Correlation-Id"
-  val correlationContext = ThreadLocalCorrelationContext.create()
   after {
     MDC.clear()
   }
 
   val defaultFilter = new CorrelationIdFilter()
-  val customFilter = new CorrelationIdFilter("customHeader", "customMdc", correlationContext)
+  val customFilter = new CorrelationIdFilter("customHeader", "customMdc")
 
   describe("CorrelationIdFilter") {
     describe("with default id names") {
-      it should behave like correlationIdFilter(defaultFilter, defaultHeaderName, defaultMdcName, correlationContext)
+      it should behave like correlationIdFilter(defaultFilter, defaultHeaderName, defaultMdcName)
     }
 
     describe("with custom id names") {
-      it should behave like correlationIdFilter(customFilter, "customHeader", "customMdc", correlationContext)
+      it should behave like correlationIdFilter(customFilter, "customHeader", "customMdc")
     }
   }
 }